# Conflicts:
#	src/main/resources/mapper/ReplyMapper.xml

+++ src/main/java/com/example/lit/mapper/user/AlertMapper.java
@@ -15,9 +15,9 @@ public interface AlertMapper {
     // 로그인한 유저 번호로 알림리스트 가져오기
     public List<AlertDTO> alertList(Long userNumber);
 
-    public Long getAlertNumber(AlertVO alertVO);
+    public int getAlertNumber(AlertVO alertVO);
 
     // 좋아요 취소시 또는 팔로우 취소시 삭제
-    public void remove(Long alertNumber);
+    public void remove(int alertNumber);
 
 }
